Update basic nestjs test application
