+++ dev-packages/e2e-tests/test-applications/nestjs-basic/.npmrc
