+++ dev-packages/e2e-tests/test-applications/nestjs-basic/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nestjs",
+  "name": "nestjs-basic",
   "version": "0.0.1",
   "private": true,
   "scripts": {
