+++ dev-packages/e2e-tests/test-applications/nestjs-basic/src/app.service.ts
@@ -3,7 +3,6 @@ import { Cron, SchedulerRegistry } from '@nestjs/schedule';
 import * as Sentry from '@sentry/nestjs';
 import { SentryCron, SentryTraced } from '@sentry/nestjs';
 import type { MonitorConfig } from '@sentry/types';
-import { makeHttpRequest } from './utils';
 
 const monitorConfig: MonitorConfig = {
   schedule: {
@@ -13,51 +12,13 @@ const monitorConfig: MonitorConfig = {
 };
 
 @Injectable()
-export class AppService1 {
+export class AppService {
   constructor(private schedulerRegistry: SchedulerRegistry) {}
-
-  testSuccess() {
-    return { version: 'v1' };
-  }
-
-  testParam(id: string) {
-    return {
-      paramWas: id,
-    };
-  }
-
-  testInboundHeaders(headers: Record<string, string>, id: string) {
-    return {
-      headers,
-      id,
-    };
-  }
-
-  async testOutgoingHttp(id: string) {
-    const data = await makeHttpRequest(`http://localhost:3030/test-inbound-headers/${id}`);
-
-    return data;
-  }
-
-  async testOutgoingFetch(id: string) {
-    const response = await fetch(`http://localhost:3030/test-inbound-headers/${id}`);
-    const data = await response.json();
-
-    return data;
-  }
 
   testTransaction() {
     Sentry.startSpan({ name: 'test-span' }, () => {
       Sentry.startSpan({ name: 'child-span' }, () => {});
     });
-  }
-
-  async testError() {
-    const exceptionId = Sentry.captureException(new Error('This is an error'));
-
-    await Sentry.flush(2000);
-
-    return { exceptionId };
   }
 
   testException(id: string) {
@@ -66,26 +27,6 @@ export class AppService1 {
 
   testExpectedException(id: string) {
     throw new HttpException(`This is an expected exception with id ${id}`, HttpStatus.FORBIDDEN);
-  }
-
-  async testOutgoingFetchExternalAllowed() {
-    const fetchResponse = await fetch('http://localhost:3040/external-allowed');
-
-    return fetchResponse.json();
-  }
-
-  async testOutgoingFetchExternalDisallowed() {
-    const fetchResponse = await fetch('http://localhost:3040/external-disallowed');
-
-    return fetchResponse.json();
-  }
-
-  async testOutgoingHttpExternalAllowed() {
-    return makeHttpRequest('http://localhost:3040/external-allowed');
-  }
-
-  async testOutgoingHttpExternalDisallowed() {
-    return makeHttpRequest('http://localhost:3040/external-disallowed');
   }
 
   @SentryTraced('wait and return a string')
@@ -122,22 +63,5 @@ export class AppService1 {
 
   async killTestCron() {
     this.schedulerRegistry.deleteCronJob('test-cron-job');
-  }
-}
-
-@Injectable()
-export class AppService2 {
-  externalAllowed(headers: Record<string, string>) {
-    return {
-      headers,
-      route: 'external-allowed',
-    };
-  }
-
-  externalDisallowed(headers: Record<string, string>) {
-    return {
-      headers,
-      route: 'external-disallowed',
-    };
   }
 }
