+++ dev-packages/e2e-tests/test-applications/nestjs-basic/src/instrument.ts
@@ -5,5 +5,4 @@ Sentry.init({
   dsn: process.env.E2E_TEST_DSN,
   tunnel: `http://localhost:3031/`, // proxy server
   tracesSampleRate: 1,
-  tracePropagationTargets: ['http://localhost:3030', '/external-allowed'],
 });
