+++ dev-packages/e2e-tests/test-applications/nestjs-basic/src/main.ts
@@ -1,0 +1,20 @@
+// Import this first
+import './instrument';
+
+// Import other modules
+import { BaseExceptionFilter, HttpAdapterHost, NestFactory } from '@nestjs/core';
+import * as Sentry from '@sentry/nestjs';
+import { AppModule } from './app.module';
+
+const app1Port = 3030;
+
+async function bootstrap() {
+  const app = await NestFactory.create(AppModule);
+
+  const { httpAdapter } = app.get(HttpAdapterHost);
+  Sentry.setupNestErrorHandler(app, new BaseExceptionFilter(httpAdapter));
+
+  await app.listen(app1Port);
+}
+
+bootstrap();
