+++ dev-packages/e2e-tests/test-applications/nestjs-basic/start-event-proxy.mjs
