+++ dev-packages/e2e-tests/test-applications/nestjs-basic/tests/errors.test.ts
@@ -53,32 +53,3 @@ test('Does not send expected exception to Sentry', async ({ baseURL }) => {
 
   expect(errorEventOccurred).toBe(false);
 });
-
-test('Does not handle expected exception if exception is thrown in module', async ({ baseURL }) => {
-  const errorEventPromise = waitForError('nestjs', event => {
-    return !event.type && event.exception?.values?.[0]?.value === 'Something went wrong in the test module!';
-  });
-
-  const response = await fetch(`${baseURL}/test-module`);
-  expect(response.status).toBe(500); // should be 400
-
-  // should never arrive, but does because the exception is not handled properly
-  const errorEvent = await errorEventPromise;
-
-  expect(errorEvent.exception?.values).toHaveLength(1);
-  expect(errorEvent.exception?.values?.[0]?.value).toBe('Something went wrong in the test module!');
-
-  expect(errorEvent.request).toEqual({
-    method: 'GET',
-    cookies: {},
-    headers: expect.any(Object),
-    url: 'http://localhost:3030/test-module',
-  });
-
-  expect(errorEvent.transaction).toEqual('GET /test-module');
-
-  expect(errorEvent.contexts?.trace).toEqual({
-    trace_id: expect.any(String),
-    span_id: expect.any(String),
-  });
-});
