--- dev-packages/e2e-tests/test-applications/nestjs/src/test-module/test.controller.ts
@@ -1,12 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { TestException } from './test.exception';
-
-@Controller('test-module')
-export class TestController {
-  constructor() {}
-
-  @Get()
-  getTest(): string {
-    throw new TestException();
-  }
-}
