--- dev-packages/e2e-tests/test-applications/nestjs/src/test-module/test.exception.ts
@@ -1,5 +1,0 @@
-export class TestException extends Error {
-  constructor() {
-    super('Something went wrong in the test module!');
-  }
-}
