--- dev-packages/e2e-tests/test-applications/nestjs/src/test-module/test.filter.ts
@@ -1,13 +1,0 @@
-import { ArgumentsHost, BadRequestException, Catch } from '@nestjs/common';
-import { BaseExceptionFilter } from '@nestjs/core';
-import { TestException } from './test.exception';
-
-@Catch(TestException)
-export class TestExceptionFilter extends BaseExceptionFilter {
-  catch(exception: unknown, host: ArgumentsHost) {
-    if (exception instanceof TestException) {
-      return super.catch(new BadRequestException(exception.message), host);
-    }
-    return super.catch(exception, host);
-  }
-}
