--- dev-packages/e2e-tests/test-applications/nestjs/src/utils.ts
@@ -1,26 +1,0 @@
-import * as http from 'http';
-
-export function makeHttpRequest(url) {
-  return new Promise(resolve => {
-    const data = [];
-
-    http
-      .request(url, httpRes => {
-        httpRes.on('data', chunk => {
-          data.push(chunk);
-        });
-        httpRes.on('error', error => {
-          resolve({ error: error.message, url });
-        });
-        httpRes.on('end', () => {
-          try {
-            const json = JSON.parse(Buffer.concat(data).toString());
-            resolve(json);
-          } catch {
-            resolve({ data: Buffer.concat(data).toString(), url });
-          }
-        });
-      })
-      .end();
-  });
-}
